Moved legacy NIDAQ controllers int main nidaq folder and updated qt3scan references, additional testing required.
