+++ src/qdlutils/applications/qt3scan/config_files/princeton_spectrometer.yaml
@@ -12,7 +12,7 @@ QT3Scan:
       experiment_name: "ZnOTrueShinySide"
 
   PositionController:
-    import_path : qdlutils.hardware.nidaq.analogoutputs.nidaqpiezocontroller
+    import_path : qdlutils.hardware.nidaq.nidaqpiezocontroller
     class_name  : QT3ScanNIDAQPositionController
     configure :
       daq_name : Dev1  # NI DAQ Device Name
