+++ src/qdlutils/applications/qt3scan/config_files/qt3scan_base.yaml
@@ -12,7 +12,7 @@ QT3Scan:
       signal_counter : ctr2  # NI DAQ counter to use for counting the input signal, e.g. ctr0, ctr1, ctr2, or ctr3
 
   PositionController:
-    import_path : qdlutils.hardware.nidaq.analogoutputs.nidaqpiezocontroller    
+    import_path : qdlutils.hardware.nidaq.nidaqpiezocontroller    
     class_name  : QT3ScanNIDAQPositionController
     configure : 
       daq_name : Dev1  # NI DAQ Device Name
